+++ cleanbcdx-builtenvironment/scripts/public-additional-scripts.js
@@ -1,0 +1,6 @@
+import './public/definitions';
+import './public/bb/accessibility';
+import './public/bh/accessibility';
+import './public/bh/rebates-archive';
+import './public/bh/definitions-form-resubmit';
+import './public/bh/tablesPattern';