+++ cleanbcdx-cleanbc/scripts/public-additional-scripts.js
@@ -1,0 +1,1 @@
+import './public/cbc/accessibility';