+++ cleanbcdx-misc/CHANGELOG.md
@@ -1,6 +1,10 @@
 ## 1.2.10 April 30, 2026
 – Added guard so script-based PDF size checks do not count
 – Temporay ability to zero out hit counts
+– Added main plugin dependency
 
 ## 1.2.9 April 30, 2026
 – Added CleanBC DX menu to WP admin bar to mimic multisite "My sites" feature
+
+## 1.2.8 - Safe Redirect Manager hit count
+## 1.2.7 - Forked from DIGIMOD-misc